refactor DualScaleSlider into a layout
