+++ battle_map_tv/window_gui.py
@@ -12,7 +12,6 @@ from PySide6.QtWidgets import (
 
 from battle_map_tv.area_of_effect import area_of_effect_shapes_to_class
 from battle_map_tv.events import EventKeys, global_event_dispatcher
-from battle_map_tv.grid import GridOverlayColor
 from battle_map_tv.layouts.app_controls import AppControlsLayout
 from battle_map_tv.layouts.area_of_effect import ColorSelectionWindow
 from battle_map_tv.layouts.base import FixedRowGridLayout
@@ -20,7 +19,6 @@ from battle_map_tv.layouts.grid_controls import GridControls
 from battle_map_tv.layouts.image_controls import ImageButtonsLayout, ImageScaleSlidersLayout
 from battle_map_tv.widgets import get_window_icon
 from battle_map_tv.widgets.buttons import StyledButton
-from battle_map_tv.widgets.sliders import DualScaleSlider, StyledSlider
 from battle_map_tv.widgets.text_based import StyledTextEdit
 from battle_map_tv.window_image import ImageWindow
